商品添加核销人已基本测试通过 goods_id全部由Integer 改成了Long 后面需要注意跟进测试

+++ src/com/yq/controller/AddressCtrl.java
@@ -97,7 +97,7 @@ public class AddressCtrl extends StringUtil{
 	}
 	
 	@RequestMapping(value = "/page/addrListTwo.html")
-	public ModelAndView listTwo(Integer goods_id,Integer goods_num,Integer addr_id,Integer cps_id,String oppen_id,HttpSession session) {
+	public ModelAndView listTwo(Long goods_id,Integer goods_num,Integer addr_id,Integer cps_id,String oppen_id,HttpSession session) {
 		address.setOppen_id(getOppen_id(session));
 		List<Address> list = addressService.list(address);
 		ModelAndView ml = new ModelAndView();
